Add files via upload
if (a > b && a>= c): Checks if a is greater than or equal to both b and c. If true, a is the largest.

else if (b >= a && b >= c): If the first condition is false, this checks if b is greater than or equal to both a and c. If true, b is the largest.

• else: If neither of the above conditions is true, then c must be the largest, so this block is executed.